Using wiki logos instead of favicons for carousel
Limited to 92 logos for bandwidth. Gathered top EN wikis by article count and pulled logos. Some logos not included due to error in programmatically pulling stats/logos, or logo was too wide. Logos are put in randomized order.

Can still improve the carousel to work better at all screen widths. May also want to load the images in with JS to be in random order each time. May come up with a better way to select featured logos in the future (e.g. logos that can be easily read or of the most popular IPs).